--- database/__init__.py
@@ -1,4 +1,0 @@
-from database._tile import TileDatabase
-from database._world import WorldDatabase
-from database._base import Database
-from database._queries import World, Tile, Color
