+++ engine/handbook/_commands.py
@@ -1,5 +1,5 @@
 from kivy.app import App
-from database import WorldDatabase
+from engine.database import WorldDatabase
 from models import Coords
 import logging
