+++ engine/ui/kv/load_ui.kv
@@ -34,6 +34,7 @@
     back_btn: back_btn
     MDLabel:
         text: "SAVES"
+        text_size: self.size
         halign: 'left'
         valign: 'top'
         text_color: "snow"
@@ -74,6 +75,7 @@
         theme_text_color: "Custom"
         text_color: "snow"
         font_style: "monocraft-ui"
+        text_size: self.size
         halign: 'left'
         valign: 'top'
         role: 'medium'
@@ -97,6 +99,7 @@
     options: options
     MDLabel:
         text: "WORLD DETAILS"
+        text_size: self.size
         halign: 'center'
         valign: 'top'
         text_color: "snow"
