+++ engine/ui/kv/main_menu.kv
@@ -6,6 +6,7 @@
 
 <TextworldMMTitle>:
     text: "TEXTWORLD"
+    text_size: self.size
     halign: 'left'
     valign: 'center'
     text_color: "snow"
