+++ engine/ui/kv/new_game_ui.kv
@@ -4,6 +4,7 @@
     text: "World Generation Settings"
     halign: 'center'
     valign: 'center'
+    text_size: self.size
     text_color: "snow"
     font_style: "monocraft-ui"
     role: 'large'    
@@ -15,6 +16,9 @@
     save_name: save_name
     MDLabel:
         text: "Save Name"
+        halign: 'center'
+        valign: 'center'
+        text_size: self.size
         text_color: "snow"
         font_style: "monocraft-ui"
         role: 'medium'
@@ -31,6 +35,9 @@
     chunk_size: chunk_size
     MDLabel:
         text: "Chunk Size"
+        halign: 'center'
+        valign: 'center'
+        text_size: self.size
         text_color: "snow"
         font_style: "monocraft-ui"
         role: 'medium'
@@ -47,6 +54,9 @@
     chunk_count: chunk_count
     MDLabel:
         text: "Chunk Count"
+        halign: 'center'
+        valign: 'center'
+        text_size: self.size
         text_color: "snow"
         font_style: "monocraft-ui"
         role: 'medium'
