+++ engine/ui/kv/tools/tile_builder.kv
@@ -2,6 +2,7 @@
 
 <TileBuilderTitle>:
     text: "TILE BUILDER"
+    text_size: self.size
     halign: 'center'
     valign: 'center'
     text_color: "snow"
