+++ engine/ui/kv/tools_ui.kv
@@ -2,8 +2,9 @@
 
 <TextworldTLTitle>:
     text: "TEXTWORLD TOOLS"
+    text_size: self.size
     halign: 'center'
-    valign: 'center'
+    valign: 'top'
     text_color: "snow"
     font_style: "monocraft-ui"
     role: 'large'
