--- generation/__init__.py
@@ -1,3 +1,0 @@
-from generation.generator import TextworldGenerator
-from generation.map import TextworldMap
-from generation.world import TextworldWorld
